fix: Escape external page's url using `esc_html` for ajax response [minor]
This is a temporary workaround for escaping url for display in HTML.
`esc_url` might return an empty string for invalid urls but we need the
url to edit a page. Proper fix would be fetching the page directly from
db when editting instead of relying on the data in the table.