new: Add functional test to guard against http/https cases [testing]
Prior to 1.2.4, there's only one cache file for both the http and the https
versions of a sitemap, which can make an http sitemap served with https links
and vice versa. The issues was fixed in 1.2.4 but no tests were added to guard
regression. This commit adds the missing tests.

Closes #27 [github]